:pencil2: shrink views.py
